requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3
python-multipart==0.0.12
PyJWT==2.9.0

# auth hashing
bcrypt==4.2.0

# recorder and s3
aiortc==1.9.0
aiohttp==3.10.5
boto3==1.35.28
